chore: move to generating code within files
